Split Unbox.swift into multiple files
- The main file was getting really large. Making it hard for newcomers
  to orientate themselves. Getting a grasp of how to use the framework
  as well as how it works. I can’t speak for maintenance but I could
  imagine it getting harder as the file grew.
- Only code change was to make a lot of the previous `private` symbols
  be `internal` instead. Exposing them to the whole module. A
  requirement for keeping everything in different files.
