+++ Sources/UnboxContainer.swift
@@ -1,0 +1,52 @@
+/**
+ *  Unbox - the easy to use Swift JSON decoder
+ *
+ *  For usage, see documentation of the classes/symbols listed in this file, as well
+ *  as the guide available at: github.com/johnsundell/unbox
+ *
+ *  Copyright (c) 2015 John Sundell. Licensed under the MIT license, as follows:
+ *
+ *  Permission is hereby granted, free of charge, to any person obtaining a copy
+ *  of this software and associated documentation files (the "Software"), to deal
+ *  in the Software without restriction, including without limitation the rights
+ *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *  copies of the Software, and to permit persons to whom the Software is
+ *  furnished to do so, subject to the following conditions:
+ *
+ *  The above copyright notice and this permission notice shall be included in all
+ *  copies or substantial portions of the Software.
+ *
+ *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ *  SOFTWARE.
+ */
+
+internal struct UnboxContainer<T: Unboxable>: UnboxableWithContext {
+    let model: T
+
+    init(unboxer: Unboxer, context: UnboxPath) throws {
+        switch context {
+        case .key(let key):
+            self.model = try unboxer.unbox(key: key)
+        case .keyPath(let keyPath):
+            self.model = try unboxer.unbox(keyPath: keyPath)
+        }
+    }
+}
+
+internal struct UnboxArrayContainer<T: Unboxable>: UnboxableWithContext {
+    let models: [T]
+
+    init(unboxer: Unboxer, context: UnboxPath) throws {
+        switch context {
+        case .key(let key):
+            self.models = try unboxer.unbox(key: key)
+        case .keyPath(let keyPath):
+            self.models = try unboxer.unbox(keyPath: keyPath)
+        }
+    }
+}
