+++ Sources/Unboxable.swift
@@ -1,0 +1,131 @@
+/**
+ *  Unbox - the easy to use Swift JSON decoder
+ *
+ *  For usage, see documentation of the classes/symbols listed in this file, as well
+ *  as the guide available at: github.com/johnsundell/unbox
+ *
+ *  Copyright (c) 2015 John Sundell. Licensed under the MIT license, as follows:
+ *
+ *  Permission is hereby granted, free of charge, to any person obtaining a copy
+ *  of this software and associated documentation files (the "Software"), to deal
+ *  in the Software without restriction, including without limitation the rights
+ *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *  copies of the Software, and to permit persons to whom the Software is
+ *  furnished to do so, subject to the following conditions:
+ *
+ *  The above copyright notice and this permission notice shall be included in all
+ *  copies or substantial portions of the Software.
+ *
+ *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ *  SOFTWARE.
+ */
+
+import Foundation
+
+/// Type alias defining what type of Dictionary that is Unboxable (valid JSON)
+public typealias UnboxableDictionary = [String : Any]
+
+/// Protocol used to declare a model as being Unboxable, for use with the unbox() function
+public protocol Unboxable {
+    /// Initialize an instance of this model by unboxing a dictionary using an Unboxer
+    init(unboxer: Unboxer) throws
+}
+
+/// Protocol used to declare a model as being Unboxable with a certain context, for use with the unbox(context:) function
+public protocol UnboxableWithContext {
+    /// The type of the contextual object that this model requires when unboxed
+    associatedtype UnboxContext
+
+    /// Initialize an instance of this model by unboxing a dictionary & using a context
+    init(unboxer: Unboxer, context: UnboxContext) throws
+}
+
+/// Protocol that types that can be used in an unboxing process must conform to. You don't conform to this protocol yourself.
+public protocol UnboxCompatible {
+    /// Unbox a value, or either throw or return nil if unboxing couldn't be performed
+    static func unbox(value: Any, allowInvalidCollectionElements: Bool) throws -> Self?
+}
+
+/// Protocol used to enable a raw type for Unboxing. See default implementations further down.
+public protocol UnboxableRawType: UnboxCompatible {
+    /// Transform an instance of this type from an unboxed number
+    static func transform(unboxedNumber: NSNumber) -> Self?
+    /// Transform an instance of this type from an unboxed string
+    static func transform(unboxedString: String) -> Self?
+}
+
+/// Protocol used to enable collections to be unboxed. Default implementations exist for Array & Dictionary
+public protocol UnboxableCollection: Collection, UnboxCompatible {
+    /// The raw collection type that this type can be unboxed from
+    associatedtype UnboxRawCollection: Collection
+    /// The value type that this collection contains
+    associatedtype UnboxValue
+
+    /// Unbox a collection, optionally allowing invalid elements & using a transform
+    static func unbox(collection: UnboxRawCollection, allowInvalidElements: Bool, transform: UnboxTransform<UnboxValue>?) throws -> Self?
+}
+
+/// Protocol used to enable an enum to be directly unboxable
+public protocol UnboxableEnum: RawRepresentable, UnboxCompatible {}
+
+/// Protocol used to enable any type to be transformed from a JSON key into a dictionary key
+public protocol UnboxableKey {
+    /// Transform an unboxed key into a key that will be used in an unboxed dictionary
+    static func transform(unboxedKey: String) -> Self?
+}
+
+/// Protocol used to enable any type as being unboxable, by transforming a raw value
+public protocol UnboxableByTransform: UnboxCompatible {
+    /// The type of raw value that this type can be transformed from. Must be a valid JSON type.
+    associatedtype UnboxRawValue
+
+    /// Attempt to transform a raw unboxed value into an instance of this type
+    static func transform(unboxedValue: UnboxRawValue) -> Self?
+}
+
+// MARK: - Default protocol implementations
+
+public extension UnboxableRawType {
+    static func unbox(value: Any, allowInvalidCollectionElements: Bool) throws -> Self? {
+        if let matchedValue = value as? Self {
+            return matchedValue
+        }
+
+        if let string = value as? String {
+            return self.transform(unboxedString: string)
+        }
+
+        if let number = value as? NSNumber {
+            return self.transform(unboxedNumber: number)
+        }
+
+        return nil
+    }
+}
+
+public extension UnboxableCollection {
+    static func unbox(value: Any, allowInvalidCollectionElements: Bool) throws -> Self? {
+        return try (value as? UnboxRawCollection).map {
+            try self.unbox(collection: $0,
+                           allowInvalidElements: allowInvalidCollectionElements,
+                           transform: nil)
+        }
+    }
+}
+
+public extension UnboxableByTransform {
+    static func unbox(value: Any, allowInvalidCollectionElements: Bool) throws -> Self? {
+        return (value as? UnboxRawValue).map(self.transform)
+    }
+}
+
+public extension UnboxableEnum {
+    static func unbox(value: Any, allowInvalidCollectionElements: Bool) throws -> Self? {
+        return (value as? RawValue).map(self.init)
+    }
+}
